+++ OpenXmlPowerTools.Tests/PowerPoint/PresentationBuilderTests.cs
@@ -38,7 +38,7 @@ namespace OxPt
                 new SlideSource(new PmlDocument(source2Pptx.FullName), 0, true),
             };
             var processedDestPptx = new FileInfo(Path.Combine(TestUtil.TempDir.FullName, "PB001-Formatting.pptx"));
-            PresentationBuilder.BuildPresentation(sources, processedDestPptx.FullName);
+            PresentationBuilder.BuildPresentation(sources).SaveAs(processedDestPptx.FullName);
         }
 
         [Fact]
@@ -54,7 +54,7 @@ namespace OxPt
                 new SlideSource(new PmlDocument(source2Pptx.FullName), 0, true),
             };
             var processedDestPptx = new FileInfo(Path.Combine(TestUtil.TempDir.FullName, "PB002-Formatting.pptx"));
-            PresentationBuilder.BuildPresentation(sources, processedDestPptx.FullName);
+            PresentationBuilder.BuildPresentation(sources).SaveAs(processedDestPptx.FullName);
         }
 
         [Fact]
@@ -73,7 +73,7 @@ namespace OxPt
                 new SlideSource(new PmlDocument(source2Pptx.FullName), 0, true),
             };
             var processedDestPptx = new FileInfo(Path.Combine(TestUtil.TempDir.FullName, "PB003-Formatting.pptx"));
-            PresentationBuilder.BuildPresentation(sources, processedDestPptx.FullName);
+            PresentationBuilder.BuildPresentation(sources).SaveAs(processedDestPptx.FullName);
         }
 
         [Fact]
@@ -92,7 +92,7 @@ namespace OxPt
                 new SlideSource(new PmlDocument(source1Pptx.FullName), 1, true),
             };
             var processedDestPptx = new FileInfo(Path.Combine(TestUtil.TempDir.FullName, "PB004-Formatting.pptx"));
-            PresentationBuilder.BuildPresentation(sources, processedDestPptx.FullName);
+            PresentationBuilder.BuildPresentation(sources).SaveAs(processedDestPptx.FullName);
         }
 
         [Fact]
@@ -112,7 +112,7 @@ namespace OxPt
                 new SlideSource(new PmlDocument(source2Pptx.FullName), 0, true),
             };
             var processedDestPptx = new FileInfo(Path.Combine(TestUtil.TempDir.FullName, "PB005-Formatting.pptx"));
-            PresentationBuilder.BuildPresentation(sources, processedDestPptx.FullName);
+            PresentationBuilder.BuildPresentation(sources).SaveAs(processedDestPptx.FullName);
         }
 
         [Fact(Skip="Bug in netcore 2.0 : https://github.com/OfficeDev/Open-Xml-PowerTools/pull/238#issuecomment-412375570")]
@@ -130,7 +130,7 @@ namespace OxPt
                 new SlideSource(new PmlDocument(sourcePptx.FullName), true),
             };
             var processedDestPptx = new FileInfo(Path.Combine(TestUtil.TempDir.FullName, "PB006-Videos.pptx"));
-            PresentationBuilder.BuildPresentation(sources, processedDestPptx.FullName);
+            PresentationBuilder.BuildPresentation(sources).SaveAs(processedDestPptx.FullName);
 
             var newMediaDataContentTypes = GetMediaDataContentTypes(processedDestPptx);