+++ docs/RELEASE_NOTES.md
@@ -1,2 +1,5 @@
+#### 1.1.0 - November 29, 2019
+- Added Slide Publishing API [#2](https://github.com/sergey-tihon/Clippit/pull/2)
+
 #### 1.0.0 - November 21, 2019
-- Initial release: latest version of OpenXmlPowerTools
+- Initial release: [latest version of OpenXmlPowerTools](https://github.com/EricWhiteDev/Open-Xml-PowerTools/tree/6e56a5f5cf662f3bd3da87945a5d3ed2329964ff)